OK - two ways are implemented to speedup the code but both ways fail!!! However, I keep these neat implementations in commented form for future use.
One way is based on bntri2bntri map to detect near-CAD-tris
and use different mapping algorithms for one-face-curved tets.
This makes a jump between two different maps that is always
there even for smooth geometries.

Second way is based on point2bntri map which again the jump
gets better but still exists!

So I decided to keep ONLY ONE MAP which
is hence based on the close-point-projection method
although it is slow!

Now trying GeomAPI_ProjectPointOnSurf might help which I am
going to do it now!